requirements.txt
flask
werkzeug
flask-session
psycopg2-binary
openai
